Fix rotation estimated by TLRotate in tangent space
Split out the manifold loss/gradient fix into a separate PR, and move the
Procrustes test into tests/test_optimization.py, per review.
